# docs/openapi/student.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/request-code": {
      "post": {
        "operationId": "AuthController_requestCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or unexpected fields"
          },
          "429": {
            "description": "Email or IP request limit exceeded"
          },
          "503": {
            "description": "SMTP unavailable or unconfigured"
          }
        },
        "summary": "Request a time-limited email sign-in code",
        "tags": [
          "Shared / Authentication"
        ]
      }
    },
    "/auth/verify-code": {
      "post": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email/code format or unexpected fields"
          },
          "401": {
            "description": "Incorrect, expired, consumed or locked verification code"
          },
          "429": {
            "description": "Email or IP verification limit exceeded"
          }
        },
        "summary": "Consume a one-time code and create a revocable session",
        "tags": [
          "Shared / Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Current session revoked"
          },
          "401": {
            "description": "Missing, expired or revoked access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current session",
        "tags": [
          "Shared / Authentication"
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "204": {
            "description": "All sessions for the current account revoked"
          },
          "401": {
            "description": "Missing, expired or revoked access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke every session of the current account",
        "tags": [
          "Shared / Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hodhod student API",
    "description": "Implemented endpoints only. Email sign-in and revocable Bearer sessions.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "EmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "student@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "RequestCodeResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "If delivery succeeds, a sign-in code will arrive shortly"
          }
        },
        "required": [
          "message"
        ]
      },
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "code": {
            "type": "string",
            "example": "083719",
            "description": "Six-digit emailed code; 11111 only in explicitly enabled development mode"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "student",
              "teacher",
              "admin"
            ]
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarId": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Tehran"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "displayName",
          "avatarId",
          "timezone",
          "createdAt"
        ]
      },
      "VerifyCodeResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "isNewUser": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "accessToken",
          "user",
          "isNewUser"
        ]
      },
...
    }
  }
}